[`tests`] Adds more tests + fix failing tests (#238)
* adds more tests

- refactor tests
- add enc-dec tests
- skips generate tests for non-lora adapters

* rm unneeded file

* fix tests

* fix

* more checks

* fix issue
